dom/base/test/test_meta_viewport0.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>meta viewport test</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>No &lt;meta name="viewport"&gt; tag</p>
  <script type="application/javascript;version=1.7">
    "use strict";

    SimpleTest.waitForExplicitFinish();

    let tests = [];

    tests.push(function test1() {
      SpecialPowers.pushPrefEnv(scaleRatio(1.0),
        function() {
          let info = getViewportInfo(800, 480);
          is(info.defaultZoom, 0,       "initial scale is unspecified");
          is(info.minZoom,     0,       "minumum scale defaults to the absolute minumum");
          is(info.maxZoom,     10,      "maximum scale defaults to the absolute maximum");
          is(info.width,       980,     "width is the default width");
          is(info.height,      588,     "height is proportional to displayHeight");
          is(info.autoSize,    false,   "autoSize is disabled by default");
          is(info.allowZoom,   true,    "zooming is enabled by default");

          info = getViewportInfo(490, 600);
          is(info.width,       980,     "width is still the default width");
          is(info.height,      1200,    "height is proportional to the new displayHeight");

          nextTest();
        });
    });

    tests.push(function test2() {
      SpecialPowers.pushPrefEnv(scaleRatio(1.5),
        function() {
          let info = getViewportInfo(800, 480);
          is(info.width,       980,     "width is still the default width");
          is(info.height,      588,     "height is still proportional to displayHeight");

          nextTest();
        });
    });

    function getViewportInfo(aDisplayWidth, aDisplayHeight) {
      let defaultZoom = {}, allowZoom = {}, minZoom = {}, maxZoom = {},
          width = {}, height = {}, autoSize = {};

      let cwu = SpecialPowers.getDOMWindowUtils(window);
      cwu.getViewportInfo(aDisplayWidth, aDisplayHeight, defaultZoom, allowZoom,
                          minZoom, maxZoom, width, height, autoSize);
      return {
        defaultZoom: defaultZoom.value,
        minZoom: minZoom.value,
        maxZoom: maxZoom.value,
        width: width.value,
        height: height.value,
        autoSize: autoSize.value,
        allowZoom: allowZoom.value
      };
    }

    function nextTest() {
      if (tests.length)
        (tests.shift())();
      else
        SimpleTest.finish();
    }
    addEventListener("load", nextTest);
  </script>
</body>
</html>
